Refactor files to a different package
